fix(FareCalculator): 추가 요금 계산 로직 수정
